+++ agentshield/checks/__init__.py
@@ -2,6 +2,7 @@ from .agent_tools import scan_agent_tools
 from .environment import scan_environment
 from .git_config import scan_git_config
 from .global_packages import scan_global_packages
+from .project import scan_project
 from .secret_files import scan_secret_files
 from .shell_history import scan_shell_history
 from .ssh import scan_ssh
@@ -11,6 +12,7 @@ __all__ = [
     "scan_agent_tools",
     "scan_git_config",
     "scan_global_packages",
+    "scan_project",
     "scan_secret_files",
     "scan_shell_history",
     "scan_ssh",